Extract Tier#case_form_json to decorator
As already done with same method for other models, as this is more of a
view concern.
