+++ app/decorators/issue_decorator.rb
@@ -1,5 +1,6 @@
 class IssueDecorator < ApplicationDecorator
   delegate_all
+  decorates_association :tiers
 
   def case_form_json
     {
